refactor/chore(nodes_registry): extract public contract into nodes_registry-sdk - add new nodes_registry-sdk crate with NodesRegistryApi, NodesRegistryError and model re-exports - node_registry layout is updated to conform standard module layour - update nodes_registry to re-export SDK types at crate root and wire modkit module client to nodes_registry_sdk - replace all crate::contract / nodes_registry::contract usages across module, domain, gateways, rest mapper and tests - NodesRegistryApi renamed to NodesRegistryClient - move DomainError -> NodesRegistryError conversion into nodes_registry domain layer - remove nodes_registry/src/contract directory - enable required chrono/uuid features in modkit-node-info for compilation - direct dependencies to plugins are removed from all docs
