+++ Cargo.toml
@@ -37,7 +37,8 @@ members = [
     "modules/system/api_gateway",
     "modules/system/grpc_hub",
     "modules/file_parser",
-    "modules/system/nodes_registry",
+    "modules/system/nodes_registry/nodes_registry",
+    "modules/system/nodes_registry/nodes_registry-sdk",
     "modules/system/module_orchestrator/module_orchestrator-sdk",
     "modules/system/module_orchestrator/module_orchestrator-grpc",
     "modules/system/module_orchestrator/module_orchestrator",
