+++ apps/hyperspot-server/Cargo.toml
@@ -40,7 +40,7 @@ hs-static-tr-plugin = { path = "../../modules/system/tenant_resolver/plugins/sta
 
 # user modules
 file_parser = { path = "../../modules/file_parser" }
-nodes_registry = { path = "../../modules/system/nodes_registry" }
+nodes_registry = { path = "../../modules/system/nodes_registry/nodes_registry" }
 
 anyhow = { workspace = true }
 tokio = { workspace = true }
