+++ modules/system/nodes_registry/nodes_registry-sdk/src/lib.rs
@@ -1,7 +1,12 @@
-pub mod client;
+#![forbid(unsafe_code)]
+#![deny(rust_2018_idioms)]
+
+pub mod api;
 pub mod error;
 
-// Re-export models from modkit-node-info
+pub use api::NodesRegistryClient;
+pub use error::NodesRegistryError;
+
 pub use modkit_node_info::{
     BatteryInfo, CpuInfo, GpuInfo, HostInfo, MemoryInfo, Node, NodeSysCap, NodeSysInfo, OsInfo,
     SysCap,
