+++ modules/system/nodes_registry/nodes_registry/src/api/rest/dto.rs
