+++ modules/system/nodes_registry/nodes_registry/src/api/rest/mod.rs
