+++ modules/system/nodes_registry/nodes_registry/src/config.rs
