+++ modules/system/nodes_registry/nodes_registry/src/domain/error.rs
@@ -39,3 +39,15 @@ impl From<modkit_node_info::NodeInfoError> for DomainError {
         }
     }
 }
+
+impl From<DomainError> for nodes_registry_sdk::NodesRegistryError {
+    fn from(e: DomainError) -> Self {
+        match e {
+            DomainError::NodeNotFound(id) => Self::NodeNotFound(id),
+            DomainError::SysInfoCollectionFailed(msg) => Self::SysInfoCollectionFailed(msg),
+            DomainError::SysCapCollectionFailed(msg) => Self::SysCapCollectionFailed(msg),
+            DomainError::InvalidInput(msg) => Self::Validation(msg),
+            DomainError::Internal(_) => Self::Internal,
+        }
+    }
+}
