+++ modules/system/nodes_registry/nodes_registry/src/domain/mod.rs
