+++ modules/system/nodes_registry/nodes_registry/src/gateways/mod.rs
